refactor: clean up tenancy, auth, runtime, and remove Supabase remnants
Tenancy & FDW:
- Rename TenantStore→TenantRepository, Storage→Repository across all call sites
- Remove dead setupFDWLegacy (~80 lines), simplify SetupFDW to 3 params
- Add 10 platform tables (without tenant_id) to FDW exclude list
- Fix MFA service direct pool bypass (m.db.Pool()→m.db.Exec, 14 sites)

Supabase removal:
- Remove supabase/supabase-demo issuers from JWT validation
- Clean ~30 Supabase references from comments, tests, and docs

Extractions:
- Extract ConnectWithRetry, EnsureDefaultTenantAndKeys, BackfillTenantIDToDefault
  from main.go (988→529 lines) into dedicated packages
- Split Execute (390-line function) into orchestrator + execute_helpers.go

Other:
- Add exponential backoff to job retry (5s·2^n capped at ~5min)
- Remove deprecated SignInAnonymous (handler, service, tests)
- Stop clearing error_message on job retry
- Document JWT scope bypass rationale and MCP/REST scope separation
