+++ cmd/fluxbase/main.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"flag"
 	"fmt"
-	"math"
 	"os"
 	"os/signal"
 	"strconv"
@@ -12,8 +11,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/google/uuid"
-	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 
@@ -23,9 +20,9 @@ import (
 	"github.com/nimbleflux/fluxbase/internal/database"
 	"github.com/nimbleflux/fluxbase/internal/database/bootstrap"
 	"github.com/nimbleflux/fluxbase/internal/database/schema"
-	"github.com/nimbleflux/fluxbase/internal/keys"
 	"github.com/nimbleflux/fluxbase/internal/migrations"
 	"github.com/nimbleflux/fluxbase/internal/storage"
+	"github.com/nimbleflux/fluxbase/internal/tenantdb"
 )
 
 var (
@@ -131,7 +128,7 @@ func main() {
 
 		// Test database connection
 		log.Info().Msg("Testing database connection...")
-		db, err := connectDatabaseWithRetry(cfg.Database, 1)
+		db, err := database.ConnectWithRetry(cfg.Database, 1)
 		if err != nil {
 			db.Close()
 			if cleanupVips != nil {
@@ -151,7 +148,7 @@ func main() {
 	}
 
 	// Initialize database connection with retry logic
-	db, err := connectDatabaseWithRetry(cfg.Database, maxRetryAttempts)
+	db, err := database.ConnectWithRetry(cfg.Database, maxRetryAttempts)
 	if err != nil {
 		if cleanupVips != nil {
 			cleanupVips()
@@ -276,14 +273,14 @@ func main() {
 
 	// Ensure default tenant and service keys exist
 	log.Info().Msg("Initializing default tenant and service keys...")
-	if err := ensureDefaultTenantAndKeys(db.Pool(), cfg); err != nil {
+	if err := tenantdb.EnsureDefaultTenantAndKeys(db.Pool(), cfg); err != nil {
 		log.Warn().Err(err).Msg("Failed to initialize default tenant and keys - continuing startup")
 	}
 
 	// Backfill NULL tenant_id to default tenant for pre-multi-tenant data.
 	// Must run after ensureDefaultTenantAndKeys because the default tenant
 	// doesn't exist during bootstrap (which runs earlier in the startup).
-	if err := backfillTenantIDToDefault(db.Pool()); err != nil {
+	if err := tenantdb.BackfillTenantIDToDefault(db.Pool()); err != nil {
 		log.Warn().Err(err).Msg("Failed to backfill tenant_id for pre-tenant data - continuing startup")
 	}
 	// Initialize tenant config loader for multi-tenant configuration overrides
@@ -302,8 +299,9 @@ func main() {
 	// Set tenant config loader for multi-tenant config overrides
 	server.SetTenantConfigLoader(tenantConfigLoader)
 
-	// Generate and set service role and anon keys for edge functions
-	// These are JWT tokens that edge functions can use to call the Fluxbase API
+	// Generate service role and anon keys for external clients (SDKs, CLI, MCP)
+	// Note: these keys are NOT passed to edge functions (blocked by runtime/env.go).
+	// Edge functions receive per-request tokens (FLUXBASE_SERVICE_TOKEN, FLUXBASE_USER_TOKEN).
 	jwtManager, err := auth.NewJWTManagerWithConfig(
 		cfg.Auth.JWTSecret,
 		cfg.Auth.JWTExpiry,
@@ -316,7 +314,7 @@ func main() {
 	}
 
 	// Get default tenant ID for JWT claims
-	defaultTenantID := getDefaultTenantID(db.Pool())
+	defaultTenantID := tenantdb.GetDefaultTenantID(db.Pool())
 
 	// Generate service role token (full admin access, bypasses RLS)
 	serviceRoleKey, err := jwtManager.GenerateServiceRoleTokenWithTenant(defaultTenantID)
@@ -440,7 +438,7 @@ func main() {
 		time.Sleep(2 * time.Second)
 		log.Warn().Msg("Force exiting - cleanup took too long")
 		os.Exit(0)
-	}()
+ 	}()
 }
 
 // getEnvInt retrieves an integer environment variable or returns a default value
@@ -453,55 +451,16 @@ func getEnvInt(key string, defaultValue int) int {
 	return defaultValue
 }
 
-// connectDatabaseWithRetry attempts to connect to the database with exponential backoff
-func connectDatabaseWithRetry(cfg config.DatabaseConfig, maxAttempts int) (*database.Connection, error) {
-	var db *database.Connection
-	var err error
-
-	for attempt := 1; attempt <= maxAttempts; attempt++ {
-		log.Info().
-			Int("attempt", attempt).
-			Int("max_attempts", maxAttempts).
-			Str("host", cfg.Host).
-			Int("port", cfg.Port).
-			Msg("Attempting to connect to database...")
-
-		db, err = database.NewConnection(cfg)
-		if err == nil {
-			log.Info().Msg("Successfully connected to database")
-			return db, nil
-		}
-
-		// If this was the last attempt, return the error
-		if attempt >= maxAttempts {
-			break
-		}
-
-		// Calculate exponential backoff (1s, 2s, 4s, 8s, 16s)
-		backoff := time.Duration(math.Pow(2, float64(attempt-1))) * time.Second
-		log.Warn().
-			Err(err).
-			Int("attempt", attempt).
-			Dur("retry_in", backoff).
-			Msg("Database connection failed, retrying...")
-		time.Sleep(backoff)
-	}
-
-	return nil, fmt.Errorf("failed to connect after %d attempts: %w", maxAttempts, err)
-}
-
 // validateStorageHealth checks if the storage provider is accessible
 func validateStorageHealth(server *api.Server) error {
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
 
-	// Access the storage service from the server
 	storageService := server.GetStorageService()
 	if storageService == nil {
 		return fmt.Errorf("storage service not initialized")
 	}
 
-	// Perform health check
 	if err := storageService.Provider.Health(ctx); err != nil {
 		return fmt.Errorf("storage health check failed: %w", err)
 	}
@@ -552,7 +511,6 @@ func printConfigSummary(cfg *config.Config) {
 	log.Info().Bool("debug_mode", cfg.Debug).Msg("  Debug Mode")
 }
 
-// getStoragePath returns the appropriate storage path/info based on provider
 func getStoragePath(storage config.StorageConfig) string {
 	if storage.Provider == "local" {
 		return storage.LocalPath
@@ -560,7 +518,6 @@ func getStoragePath(storage config.StorageConfig) string {
 	return storage.S3Bucket
 }
 
-// getEmailProviderInfo returns email provider info with masked credentials
 func getEmailProviderInfo(email config.EmailConfig) string {
 	if !email.Enabled {
 		return "disabled"
@@ -569,419 +526,4 @@ func getEmailProviderInfo(email config.EmailConfig) string {
 		return fmt.Sprintf("smtp (%s:%d)", email.SMTPHost, email.SMTPPort)
 	}
 	return email.Provider
-}
-
-// ensureDefaultTenantAndKeys ensures the default tenant and service keys exist
-func ensureDefaultTenantAndKeys(pool *pgxpool.Pool, cfg *config.Config) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	var tenantID uuid.UUID
-	var tenantExists bool
-
-	// Check if default tenant exists (with retry for transient connection errors
-	// that can occur right after pool recreation).
-	var err error
-	for attempt := 0; attempt < 3; attempt++ {
-		err = pool.QueryRow(
-			ctx,
-			"SELECT id, true FROM platform.tenants WHERE slug = 'default' AND deleted_at IS NULL",
-		).Scan(&tenantID, &tenantExists)
-		if err == nil {
-			break
-		}
-		if isNoRowsError(err) {
-			err = nil
-			break
-		}
-		// Also check for the standard pgx "no rows" message
-		if strings.Contains(err.Error(), "no rows in result set") {
-			err = nil
-			break
-		}
-		if attempt < 2 {
-			log.Warn().Err(err).Msg("Retrying default tenant check due to connection error")
-			time.Sleep(500 * time.Millisecond)
-		}
-	}
-	if err != nil && !isNoRowsError(err) {
-		return fmt.Errorf("failed to check for default tenant: %w", err)
-	}
-
-	if !tenantExists {
-		// Create default tenant
-		tenantName := cfg.Tenants.Default.Name
-		if tenantName == "" {
-			tenantName = "Default"
-		}
-
-		err := pool.QueryRow(
-			ctx,
-			"INSERT INTO platform.tenants (slug, name, is_default) VALUES ('default', $1, true) RETURNING id",
-			tenantName,
-		).Scan(&tenantID)
-		if err != nil {
-			return fmt.Errorf("failed to create default tenant: %w", err)
-		}
-		log.Info().Str("id", tenantID.String()).Msg("Created default tenant")
-	} else {
-		log.Debug().Str("id", tenantID.String()).Msg("Default tenant already exists")
-	}
-
-	// Handle service key
-	if err := ensureServiceKey(ctx, pool, cfg, tenantID, keys.KeyTypeTenantService); err != nil {
-		return fmt.Errorf("failed to ensure service key: %w", err)
-	}
-
-	// Handle anon key
-	if err := ensureServiceKey(ctx, pool, cfg, tenantID, keys.KeyTypeAnon); err != nil {
-		return fmt.Errorf("failed to ensure anon key: %w", err)
-	}
-
-	return nil
-}
-
-// ensureServiceKey ensures a service key exists for the given type
-// For database-per-tenant, keys are stored in auth.service_keys in the tenant's database
-func ensureServiceKey(ctx context.Context, pool *pgxpool.Pool, cfg *config.Config, tenantID uuid.UUID, keyType string) error {
-	var configKey string
-	var keyName string
-
-	switch keyType {
-	case keys.KeyTypeTenantService:
-		configKey = cfg.Tenants.Default.ServiceKey
-		if cfg.Tenants.Default.ServiceKeyFile != "" {
-			if data, err := os.ReadFile(cfg.Tenants.Default.ServiceKeyFile); err == nil {
-				configKey = strings.TrimSpace(string(data))
-			} else {
-				log.Warn().Err(err).Str("file", cfg.Tenants.Default.ServiceKeyFile).Msg("Failed to read service key file")
-			}
-		}
-		keyName = "Default Service Key"
-	case keys.KeyTypeAnon:
-		configKey = cfg.Tenants.Default.AnonKey
-		if cfg.Tenants.Default.AnonKeyFile != "" {
-			if data, err := os.ReadFile(cfg.Tenants.Default.AnonKeyFile); err == nil {
-				configKey = strings.TrimSpace(string(data))
-			} else {
-				log.Warn().Err(err).Str("file", cfg.Tenants.Default.AnonKeyFile).Msg("Failed to read anon key file")
-			}
-		}
-		keyName = "Default Anon Key"
-	default:
-		return fmt.Errorf("unsupported key type: %s", keyType)
-	}
-
-	// Check for existing key of this type in auth.service_keys
-	var existingKeyID uuid.UUID
-	var existingKeyHash string
-	err := pool.QueryRow(
-		ctx,
-		"SELECT id, key_hash FROM auth.service_keys WHERE key_type = $1 AND enabled = true AND revoked_at IS NULL",
-		keyType,
-	).Scan(&existingKeyID, &existingKeyHash)
-	hasExistingKey := err == nil
-
-	if configKey != "" {
-		// Config-managed key provided
-		keyHash, err := keys.HashKey(configKey)
-		if err != nil {
-			return fmt.Errorf("failed to hash config key: %w", err)
-		}
-
-		if hasExistingKey {
-			// Check if hash matches
-			if keys.VerifyKey(configKey, existingKeyHash) {
-				log.Debug().Str("type", keyType).Msg("Config-managed key already stored")
-				return nil
-			}
-			// Key changed, disable old and create new
-			_, err := pool.Exec(
-				ctx,
-				"UPDATE auth.service_keys SET enabled = false WHERE id = $1",
-				existingKeyID,
-			)
-			if err != nil {
-				return fmt.Errorf("failed to disable old key: %w", err)
-			}
-		}
-
-		// Insert new config-managed key
-		keyPrefix := keys.ExtractPrefix(configKey)
-		_, err = pool.Exec(
-			ctx,
-			`INSERT INTO auth.service_keys 
-			(name, key_hash, key_prefix, key_type, enabled, scopes, rate_limit_per_minute)
-			VALUES ($1, $2, $3, $4, true, $5, $6)`,
-			keyName, keyHash, keyPrefix, keyType, getDefaultScopes(keyType), getDefaultRateLimit(keyType),
-		)
-		if err != nil {
-			return fmt.Errorf("failed to insert config-managed key: %w", err)
-		}
-
-		log.Info().Str("type", keyType).Msg("Stored config-managed key")
-		return nil
-	}
-
-	// No config key provided
-	if hasExistingKey {
-		log.Debug().Str("type", keyType).Msg("Service key already exists")
-		return nil
-	}
-
-	// Generate new key
-	_, keyHash, keyPrefix, err := keys.GenerateKey(keyType)
-	if err != nil {
-		return fmt.Errorf("failed to generate key: %w", err)
-	}
-
-	// Insert generated key
-	_, err = pool.Exec(
-		ctx,
-		`INSERT INTO auth.service_keys 
-		(name, key_hash, key_prefix, key_type, enabled, scopes, rate_limit_per_minute)
-		VALUES ($1, $2, $3, $4, true, $5, $6)`,
-		keyName, keyHash, keyPrefix, keyType, getDefaultScopes(keyType), getDefaultRateLimit(keyType),
-	)
-	if err != nil {
-		return fmt.Errorf("failed to insert generated key: %w", err)
-	}
-
-	log.Info().
-		Str("type", keyType).
-		Str("prefix", keyPrefix).
-		Msg("Generated new service key - configure via tenants.default.anon_key or tenants.default.service_key to persist")
-
-	return nil
-}
-
-// getDefaultScopes returns default scopes for a key type
-func getDefaultScopes(keyType string) []string {
-	switch keyType {
-	case keys.KeyTypeTenantService:
-		return []string{"*"}
-	case keys.KeyTypeAnon:
-		return []string{"read"}
-	default:
-		return []string{}
-	}
-}
-
-// getDefaultRateLimit returns default rate limit for a key type
-func getDefaultRateLimit(keyType string) int {
-	switch keyType {
-	case keys.KeyTypeTenantService:
-		return 10000
-	case keys.KeyTypeAnon:
-		return 60
-	default:
-		return 60
-	}
-}
-
-// getDefaultTenantID returns the default tenant ID
-func getDefaultTenantID(pool *pgxpool.Pool) *string {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var tenantID string
-	err := pool.QueryRow(
-		ctx,
-		"SELECT id::text FROM platform.tenants WHERE slug = 'default' AND deleted_at IS NULL",
-	).Scan(&tenantID)
-	if err != nil {
-		log.Debug().Err(err).Msg("Failed to get default tenant ID")
-		return nil
-	}
-	return &tenantID
-}
-
-// isNoRowsError checks if the error is a "no rows" error
-func isNoRowsError(err error) bool {
-	return err != nil && (err.Error() == "no rows in result set" || strings.Contains(err.Error(), "no rows"))
-}
-
-// backfillTenantIDToDefault assigns NULL tenant_id rows to the default tenant.
-// This handles the upgrade path from pre-multi-tenant Fluxbase where all data
-// was created without tenant context. Without this backfill, selecting the
-// default tenant in the admin UI makes all pre-existing data invisible because
-// has_tenant_access(NULL) returns FALSE when a tenant context is active.
-//
-// Tables where NULL is semantically meaningful (settings cascade, shared OAuth
-// providers, global service keys) are excluded from the backfill.
-func backfillTenantIDToDefault(pool *pgxpool.Pool) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-	defer cancel()
-
-	// Get default tenant ID
-	var defaultTenantID string
-	err := pool.QueryRow(
-		ctx,
-		"SELECT id::text FROM platform.tenants WHERE is_default = true AND deleted_at IS NULL LIMIT 1",
-	).Scan(&defaultTenantID)
-	if err != nil {
-		if err.Error() == "no rows in result set" || strings.Contains(err.Error(), "no rows") {
-			log.Debug().Msg("No default tenant found, skipping tenant_id backfill")
-			return nil
-		}
-		return fmt.Errorf("failed to get default tenant: %w", err)
-	}
-
-	// Tables to backfill: user data that should belong to the default tenant.
-	// Excluded:
-	//   - platform.instance_settings (NULL = instance defaults shared with all tenants)
-	//   - platform.oauth_providers (NULL = shared SSO providers)
-	//   - platform.service_keys (NULL = global_service keys)
-	//   - platform.enabled_extensions (NULL = instance-level extensions)
-	//   - auth.service_keys (NULL = global_service keys)
-	//   - auth.saml_providers (NULL = shared SAML providers, like OAuth)
-	tenantIDDedupTables := map[string][]string{
-		"auth.users":               {"email"},
-		"functions.edge_functions": {"name, namespace"},
-		"storage.buckets":          {"name"},
-		"branching.branches":       {"name", "slug"},
-		"branching.github_config":  {"repository"},
-		"mcp.custom_tools":         {"name, namespace"},
-		"mcp.custom_resources":     {"uri, namespace"},
-	}
-
-	tables := []string{
-		"auth.users",
-		"auth.client_keys",
-		"auth.impersonation_sessions",
-		"auth.webhooks",
-		"auth.webhook_deliveries",
-		"auth.webhook_events",
-		"auth.saml_providers",
-		"auth.sessions",
-		"auth.oauth_links",
-		"auth.oauth_tokens",
-		"auth.mfa_factors",
-		"auth.saml_sessions",
-		"auth.magic_links",
-		"auth.otp_codes",
-		"auth.email_verification_tokens",
-		"auth.password_reset_tokens",
-		"auth.two_factor_setups",
-		"auth.two_factor_recovery_attempts",
-		"auth.oauth_logout_states",
-		"auth.mcp_oauth_clients",
-		"auth.mcp_oauth_codes",
-		"auth.mcp_oauth_tokens",
-		"auth.client_key_usage",
-		"auth.service_key_revocations",
-		"functions.edge_functions",
-		"functions.edge_executions",
-		"functions.edge_files",
-		"functions.edge_triggers",
-		"functions.secrets",
-		"functions.secret_versions",
-		"functions.shared_modules",
-		"functions.function_dependencies",
-		"jobs.functions",
-		"jobs.function_files",
-		"jobs.workers",
-		"jobs.queue",
-		"ai.knowledge_bases",
-		"ai.knowledge_base_permissions",
-		"ai.documents",
-		"ai.document_permissions",
-		"ai.chunks",
-		"ai.entities",
-		"ai.document_entities",
-		"ai.entity_relationships",
-		"ai.providers",
-		"ai.chatbots",
-		"ai.chatbot_knowledge_bases",
-		"ai.conversations",
-		"ai.messages",
-		"ai.query_audit_log",
-		"ai.retrieval_log",
-		"ai.table_export_sync_configs",
-		"ai.user_chatbot_usage",
-		"ai.user_provider_preferences",
-		"ai.user_quotas",
-		"rpc.procedures",
-		"rpc.executions",
-		"realtime.schema_registry",
-		"storage.buckets",
-		"storage.objects",
-		"storage.chunked_upload_sessions",
-		"storage.object_permissions",
-		"branching.branches",
-		"branching.activity_log",
-		"branching.branch_access",
-		"branching.github_config",
-		"branching.migration_history",
-		"branching.seed_execution_log",
-		"logging.entries",
-		"logging.entries_ai",
-		"logging.entries_custom",
-		"logging.entries_execution",
-		"logging.entries_http",
-		"logging.entries_security",
-		"logging.entries_system",
-		"mcp.custom_resources",
-		"mcp.custom_tools",
-		"platform.invitation_tokens",
-	}
-
-	var totalBackfilled int
-	for _, table := range tables {
-		if dedupSets, needsDedup := tenantIDDedupTables[table]; needsDedup {
-			for _, dedupCols := range dedupSets {
-				dedupQuery := fmt.Sprintf(
-					"DELETE FROM %s WHERE id IN (SELECT id FROM (SELECT id, row_number() OVER (PARTITION BY %s ORDER BY created_at DESC) AS rn FROM %s WHERE tenant_id IS NULL) sub WHERE rn > 1)",
-					table, dedupCols, table,
-				)
-				if dedupResult, err := pool.Exec(ctx, dedupQuery); err != nil {
-					log.Warn().Err(err).Str("table", table).Str("cols", dedupCols).Msg("Failed to dedup NULL-tenant rows before backfill")
-				} else if n := dedupResult.RowsAffected(); n > 0 {
-					log.Info().Str("table", table).Str("cols", dedupCols).Int64("duplicates_removed", n).Msg("Removed duplicate NULL-tenant rows before backfill")
-				}
-
-				conflictQuery := fmt.Sprintf(
-					"DELETE FROM %s WHERE id IN (SELECT n.id FROM %s n WHERE n.tenant_id IS NULL AND EXISTS (SELECT 1 FROM %s e WHERE e.tenant_id = $1 AND (%s)))",
-					table, table, table, buildJoinCondition(dedupCols, "n", "e"),
-				)
-				if conflictResult, err := pool.Exec(ctx, conflictQuery, defaultTenantID); err != nil {
-					log.Warn().Err(err).Str("table", table).Str("cols", dedupCols).Msg("Failed to remove NULL-tenant rows conflicting with existing tenant rows")
-				} else if n := conflictResult.RowsAffected(); n > 0 {
-					log.Info().Str("table", table).Str("cols", dedupCols).Int64("conflicts_removed", n).Msg("Removed NULL-tenant rows that would conflict with existing tenant rows")
-				}
-			}
-		}
-
-		result, err := pool.Exec(
-			ctx,
-			fmt.Sprintf("UPDATE %s SET tenant_id = $1::uuid WHERE tenant_id IS NULL", table),
-			defaultTenantID,
-		)
-		if err != nil {
-			log.Warn().Err(err).Str("table", table).Msg("Failed to backfill tenant_id")
-			continue
-		}
-		if n := result.RowsAffected(); n > 0 {
-			log.Info().Str("table", table).Int64("rows", n).Msg("Backfilled tenant_id to default tenant")
-			totalBackfilled += int(n)
-		}
-	}
-
-	if totalBackfilled > 0 {
-		log.Info().Int("total_rows", totalBackfilled).Msg("Tenant_id backfill complete")
-	} else {
-		log.Debug().Msg("No NULL tenant_id rows found to backfill")
-	}
-
-	return nil
-}
-
-func buildJoinCondition(cols, leftAlias, rightAlias string) string {
-	parts := strings.Split(cols, ",")
-	conditions := make([]string, len(parts))
-	for i, col := range parts {
-		col = strings.TrimSpace(col)
-		conditions[i] = fmt.Sprintf("%s.%s = %s.%s", leftAlias, col, rightAlias, col)
-	}
-	return strings.Join(conditions, " AND ")
 }
