+++ internal/auth/service_tokens.go
@@ -22,7 +22,7 @@ func (s *Service) ValidateTokenWithSecret(token, secretKey string) (*TokenClaims
 }
 
 // ValidateServiceRoleToken validates a JWT containing a role claim (anon, service_role, authenticated)
-// This is used for Supabase-compatible client keys which are JWTs with role claims.
+// This is used for client keys which are JWTs with role claims.
 // Unlike user tokens, these don't require user lookup or revocation checks.
 func (s *Service) ValidateServiceRoleToken(token string) (*TokenClaims, error) {
 	return s.jwtManager.ValidateServiceRoleToken(token)
